--- a/mpf/benchmarks/machine_files/shows/shows/minimal_light_show_token.yaml
+++ b/mpf/benchmarks/machine_files/shows/shows/minimal_light_show_token.yaml
@@ -1,4 +1,4 @@
 #show_version=5
 - duration: .001
   lights:
-    (tag): (color)
+    (led): red