--- a/mpf/benchmarks/test_benchmark_light_shows.py
+++ b/mpf/benchmarks/test_benchmark_light_shows.py
@@ -48,13 +48,17 @@ class BenchmarkLightShows(MpfGameTestCase):
     def testBenchmark(self):
         baseline = self._benchmark(partial(self._event_and_run, "random_event", "random_event2"), "baseline")
         minimal_show = self._benchmark(partial(self._event_and_run, "play_minimal_light_show", "stop_minimal_light_show"), "minimal_show")
+        minimal_show_token = self._benchmark(partial(self._event_and_run, "play_minimal_light_show_token", "stop_minimal_light_show_token"), "minimal_show_token")
         all_leds = self._benchmark(partial(self._event_and_run, "play_single_step_tag_playfield", "stop_single_step_tag_playfield"), "all_leds_tag")
         multi_step = self._benchmark(partial(self._event_and_run, "play_multi_step", "stop_multi_step"), "multi_step", num=500)
 
-        print("Baseline: {:.5f}ms One LED: +{:.5f}ms {:.2f} fps; 30 LEDs: +{:.5f}ms {:.2f} fps; "
+        print("Baseline: {:.5f}ms One LED: +{:.5f}ms {:.2f} fps; "
+              "One LED (token): +{:.5f}ms {:.2f} fps; "
+              "30 LEDs: +{:.5f}ms {:.2f} fps; "
               "Multi Step: +{:.5f} {:.2f} fps;".format(
             baseline * 1000,
             (minimal_show - baseline) * 1000, 1 / (minimal_show - baseline),
+            (minimal_show_token - baseline) * 1000, 1 / (minimal_show_token - baseline),
             (all_leds - baseline) * 1000, 1 / (all_leds - baseline),
             (multi_step - baseline) * 1000, 1 / (multi_step - baseline)
             ))
@@ -73,6 +77,7 @@ class BenchmarkLightShows(MpfGameTestCase):
             self.advance_time_and_run(.01)
             for channel in channel_list:
                 brightness = channel.current_brightness
+            self.machine.events.post(event2)
 
         end = time.time()
         self.advance_time_and_run()