… mpf/benchmarks/machine_files/shows/config/config.yaml
… mpf/tests/loop.py
… mpf/tests/machine_files/shows/config/test_shows.yaml