Benchmarks (old/new):

Python 3.6.7:
Baseline: 0.06306ms One LED: +0.46666ms 2142.88 fps; One LED (token): +0.50855ms 1966.38 fps; 30 LEDs: +2.42702ms 412.03 fps; Multi Step: +2.38754 418.84 fps;
Baseline: 0.06545ms One LED: +0.35594ms 2809.47 fps; One LED (token): +0.35936ms 2782.72 fps; 30 LEDs: +1.98664ms 503.36 fps; Multi Step: +1.69642 589.48 fps;

PYPY 6.0.0:
Baseline: 0.00615ms One LED: +0.05008ms 19968.83 fps; One LED (token): +0.05684ms 17592.36 fps; 30 LEDs: +0.20977ms 4767.07 fps; Multi Step: +0.28906 3459.48 fps;
Baseline: 0.00611ms One LED: +0.03172ms 31530.11 fps; One LED (token): +0.03393ms 29473.34 fps; 30 LEDs: +0.18480ms 5411.28 fps; Multi Step: +0.16721 5980.64 fps;

PYPY 7.0.0:
Baseline: 0.00604ms One LED: +0.05009ms 19965.37 fps; One LED (token): +0.05545ms 18034.74 fps; 30 LEDs: +0.22030ms 4539.21 fps; Multi Step: +0.28887 3461.77 fps;
Baseline: 0.00606ms One LED: +0.03441ms 29058.21 fps; One LED (token): +0.03454ms 28951.20 fps; 30 LEDs: +0.18447ms 5421.00 fps; Multi Step: +0.19171 5216.10 fps;