@jabdoa2 2019-04-26
http://docs.missionpinball.org/en/dev/finalization/software.html#run-mpf-with-pypy